requirements.txt
# Original YaTC stack. Use Python 3.8 or 3.9, not Python 3.10+.
numpy==1.19.5
torch==1.9.0
torchvision==0.10.0
timm==0.3.2
scikit-learn==0.24.2
scikit-image==0.19.3
matplotlib==3.3.4
Pillow
tensorboard
scapy
tqdm
